chore: split diagnostics.rs test modules for the 2000-line gate (#8889)
